Starts fixing #127
- moved updateStat to new file
- Added verify check to logic, checking if file was updated
