+++ setup/step2.php
@@ -82,40 +82,15 @@ require_once('../core/php/loadVars.php'); ?>
 		document.getElementById('settingsMenuVars').action = "../core/php/settingsSave.php";
 		document.getElementById('settingsMenuVars').submit();
 	}
-	function updateStatus(status)
-	{
-		var urlForSend = './updateSetupStatus.php?format=json'
-		var data = {status: status };
-		$.ajax({
-				  url: urlForSend,
-				  dataType: 'json',
-				  data: data,
-				  type: 'POST',
-		success: function(data)
-		{
-			if(status == "finished")
-			{
-				defaultSettings();
-			}
-			else
-			{
-				customSettings();
-			}
-	  	},
-			});
-		return false;
-	}
-
-		var popupSettingsArray = JSON.parse('<?php echo json_encode($popupSettingsArray) ?>');
+	var popupSettingsArray = JSON.parse('<?php echo json_encode($popupSettingsArray) ?>');
 	var fileArray = JSON.parse('<?php echo json_encode($config['watchList']) ?>');
 	var countOfWatchList = <?php echo $i; ?>;
 	var countOfAddedFiles = 0;
 	var countOfClicks = 0;
 	var locationInsert = "newRowLocationForWatchList";
 	var logTrimType = "<?php echo $logTrimType; ?>";
-
-
 </script>
+<script src="stepsJavascript.js"></script>
 <script src="../core/js/settingsMain.js"></script>
 <?php readfile('../core/html/popup.html') ?>
 </html>
