+++ setup/welcome.php
@@ -62,28 +62,6 @@ if(file_exists($baseUrl.'conf/config.php'))
 		//change setupProcess to page1
 		document.getElementById('defaultVarsForm').submit();
 	}
-	function updateStatus(status)
-	{
-		var urlForSend = './updateSetupStatus.php?format=json'
-		var data = {status: status };
-		$.ajax({
-				  url: urlForSend,
-				  dataType: 'json',
-				  data: data,
-				  type: 'POST',
-		success: function(data)
-		{
-			if(status == "finished")
-			{
-				defaultSettings();
-			}
-			else
-			{
-				customSettings();
-			}
-	  	},
-			});
-		return false;
-	}
 </script>
+<script src="stepsJavascript.js"></script>
 </html>
